alhamdo li Allah changed typeof 'task_t.parents_id_list_size' and 'task_t.children_is_list_size' from uint8_t to uint32_t

+++ include/task_t.h
@@ -12,10 +12,10 @@ struct task_t
     uint8_t progress;
 
     /*array of parents IDs*/
-    uint8_t parents_id_list_size;
-    uint8_t children_id_list_size;
-    uint32_t* parents_id_list;
-    uint32_t* children_id_list;
+    uint32_t parents_id_list_size;
+    uint32_t children_id_list_size;
+    uint32_t *parents_id_list;
+    uint32_t *children_id_list;
 
     //time_t or equivalent
 };
